Rename JsonViewer model folder to utils for consistency.
